was written DpForLeftSectionTexts, DpForUnderFourIconsTexts and complete
