+++ src/test/java/ru/training/at/hw2/ex1/dataproviders/DpForUnderFourIconsTexts.java
@@ -14,9 +14,12 @@ public class DpForUnderFourIconsTexts {
                 {1, "To include good practices\n"
                         + "and ideas from successful\n"
                         + "EPAM project"},
+
                 {2, "To be flexible and\n"
                        + "customizable"},
+
                 {3, "To be multiplatform"},
+
                 {4, "Already have good base\n"
                        + "(about 20 internal and\n"
                        + "some external projects),\n"
